+++ algorithms/active/discrimination-tree/src/test/java/de/learnlib/algorithms/discriminationtree/DTLearnerMealyIT.java
@@ -30,6 +30,7 @@ public class DTLearnerMealyIT extends AbstractMealyLearnerIT {
 
     @Override
     protected <I, O> void addLearnerVariants(Alphabet<I> alphabet,
+                                             int targetSize,
                                              MealyMembershipOracle<I, O> mqOracle,
                                              MealyLearnerVariantList<I, O> variants) {
         DTLearnerMealyBuilder<I, O> builder = new DTLearnerMealyBuilder<>();