+++ algorithms/active/discrimination-tree/src/test/java/de/learnlib/algorithms/discriminationtree/DTLearnerMooreIT.java
@@ -30,6 +30,7 @@ public class DTLearnerMooreIT extends AbstractMooreLearnerIT {
 
     @Override
     protected <I, O> void addLearnerVariants(Alphabet<I> alphabet,
+                                             int targetSize,
                                              MooreMembershipOracle<I, O> mqOracle,
                                              MooreLearnerVariantList<I, O> variants) {
         DTLearnerMooreBuilder<I, O> builder = new DTLearnerMooreBuilder<>();