+++ algorithms/active/lstar/src/test/java/de/learnlib/algorithms/lstar/it/ExtensibleLStarMealyIT.java
@@ -32,6 +32,7 @@ public class ExtensibleLStarMealyIT extends AbstractMealyLearnerIT {
 
     @Override
     protected <I, O> void addLearnerVariants(Alphabet<I> alphabet,
+                                             int targetSize,
                                              MealyMembershipOracle<I, O> mqOracle,
                                              MealyLearnerVariantList<I, O> variants) {
         ExtensibleLStarMealyBuilder<I, O> builder = new ExtensibleLStarMealyBuilder<>();