+++ algorithms/active/oml/src/test/java/de/learnlib/algorithms/oml/lstar/mealy/it/OptimalLStarMealyIT.java
@@ -27,6 +27,7 @@ public class OptimalLStarMealyIT extends AbstractMealyLearnerIT {
 
     @Override
     protected <I, O> void addLearnerVariants(Alphabet<I> alphabet,
+                                             int targetSize,
                                              MealyMembershipOracle<I, O> mqOracle,
                                              MealyLearnerVariantList<I, O> variants) {