+++ test-support/learner-it-support/src/main/java/de/learnlib/testsupport/it/learner/AbstractMealyLearnerIT.java
@@ -24,7 +24,7 @@ import de.learnlib.driver.util.StateLocalInputMealySimulatorSUL;
 import de.learnlib.examples.LearningExample.MealyLearningExample;
 import de.learnlib.examples.LearningExample.StateLocalInputMealyLearningExample;
 import de.learnlib.examples.LearningExamples;
-import de.learnlib.oracle.equivalence.SimulatorEQOracle;
+import de.learnlib.oracle.equivalence.MealySimulatorEQOracle;
 import de.learnlib.oracle.equivalence.mealy.StateLocalInputMealySimulatorEQOracle;
 import de.learnlib.oracle.membership.SimulatorOracle.MealySimulatorOracle;
 import de.learnlib.oracle.membership.StateLocalInputSULOracle;
@@ -69,13 +69,14 @@ public abstract class AbstractMealyLearnerIT {
             MealyLearningExample<I, O> example) {
 
         final Alphabet<I> alphabet = example.getAlphabet();
-        final MealyMembershipOracle<I, O> mqOracle = new MealySimulatorOracle<>(example.getReferenceAutomaton());
+        final MealyMachine<?, I, ?, O> reference = example.getReferenceAutomaton();
+        final MealyMembershipOracle<I, O> mqOracle = new MealySimulatorOracle<>(reference);
         final MealyLearnerVariantListImpl<I, O> variants = new MealyLearnerVariantListImpl<>();
-        addLearnerVariants(alphabet, mqOracle, variants);
+        addLearnerVariants(alphabet, reference.size(), mqOracle, variants);
 
         return LearnerITUtil.createExampleITCases(example,
                                                   variants,
-                                                  new SimulatorEQOracle<>(example.getReferenceAutomaton()));
+                                                  new MealySimulatorEQOracle<>(example.getReferenceAutomaton()));
     }
 
     private <I, O> List<UniversalDeterministicLearnerITCase<I, Word<O>, MealyMachine<?, I, ?, O>>> createPartialVariantsITCase(
@@ -92,7 +93,7 @@ public abstract class AbstractMealyLearnerIT {
         final MealyMembershipOracle<I, O> mqOracle =
                 new StateLocalInputSULOracle<>(new StateLocalInputMealySimulatorSUL<>(partialRef), undefinedOutput);
         final MealyLearnerVariantListImpl<I, O> variants = new MealyLearnerVariantListImpl<>();
-        addLearnerVariants(alphabet, mqOracle, variants);
+        addLearnerVariants(alphabet, reference.size(), mqOracle, variants);
 
         final MealyEquivalenceOracle<I, O> eqOracle =
                 new StateLocalInputMealySimulatorEQOracle<>(partialRef, alphabet, undefinedOutput);
@@ -101,17 +102,20 @@ public abstract class AbstractMealyLearnerIT {
     }
 
     /**
-     * Adds, for a given setup, all the variants of the Mealy machine learner to be tested to the specified {@link
-     * LearnerVariantList variant list}.
+     * Adds, for a given setup, all the variants of the Mealy machine learner to be tested to the specified
+     * {@link LearnerVariantList variant list}.
      *
      * @param alphabet
      *         the input alphabet
+     * @param targetSize
+     *         the size of the target automaton
      * @param mqOracle
      *         the membership oracle
      * @param variants
      *         list to add the learner variants to
      */
     protected abstract <I, O> void addLearnerVariants(Alphabet<I> alphabet,
+                                                      int targetSize,
                                                       MealyMembershipOracle<I, O> mqOracle,
                                                       MealyLearnerVariantList<I, O> variants);
 }